Add automerge to hypermerge
Create the automerge document automatically when creating the
hypermerge.

For now, the tests have two extra hypermerge docs for replicated
hypercores (the automerge doc is unused). The next step is to add
hypercores for the peers to hypermerge.

Because the actor ids are now random, the unit test for the
conflict is a bit more complex.