fix errors on null values in documents
In Arthropod, creating tasks with a null assignee value was resulting in
errors due to this traversal logic. Guarding against that case here
fixes the issue.

I'm not familiar enough with the context here to know if this might be
a more generalized problem. I'm guessing it's an issue with this fork
and not the main branch of hypermerge, assuming that hypermerge has been
used with documents that contain null values before.